done with all shopping cart and order routes

+++ app/api/cart_routes.py
@@ -1,6 +1,6 @@
-from flask import Blueprint, jsonify, session, redirect, request
+from flask import Blueprint, session, request
 from app import db
-from app.models import ShoppingCart, CartItem, Product, ProductImage
+from app.models import ShoppingCart, CartItem, Product, Order, OrderItem, ProductImage
 from flask_login import current_user, login_required
 
 
@@ -20,12 +20,12 @@ def createShoppingCart():
     try:
         db.session.add(cart)
         db.session.commit()
-        return {'shopping_cart': cart.to_dict()}
+        return {'shopping_cart': cart.to_dict()}, 201
     except:
         return {'errors': {'message': 'There was an error creating shopping cart'}}
 
 
-#Get items for Shopping Cart
+#Get items for Shopping Cart, sets total of cart
 @cart_route.route('/<int:cart_id>')
 @login_required
 def getShoppingCart(cart_id):
@@ -54,6 +54,9 @@ def getShoppingCart(cart_id):
         item['images'] = [image.to_dict() for image in images]
 
 
+    #set total of cart
+    cart_by_id.total = total
+    db.session.commit()
 
     cart_dict = cart_by_id.to_dict()
     cart_dict['items'] = items_dict
@@ -61,22 +64,50 @@ def getShoppingCart(cart_id):
 
     return {'shoppingCart': cart_dict}, 200
 
-#Remove item to shopping cart
-@cart_route.route('/<int:item_id>', methods=['DELETE'])
+#Add item via ITEM id
+@cart_route.route('/add/<int:item_id>/<int:quantity>', methods=['POST'])
 @login_required
-def removeItem(item_id):
+def addItem(item_id, quantity):
 
     cart_item = db.session.query(CartItem).filter(CartItem.id == item_id).first()
 
-    try:
-        db.session.delete(cart_item)
-        db.session.commit()
-        return {'message': 'Cart item successfuly deleted'}
-    except:
-        return {'errors': {'message': 'There was an error in deleted item from cart'}}
+    if request.method == "POST":
+        cart_item.quantity += quantity
+        try:
+            db.session.commit()
+            return {'cart_item': cart_item.to_dict()}, 201
+        except:
+            return {'errors': {'message': 'There was an error in adding quantity to item'}}
 
+#Subtract/delete item via ITEM id
+@cart_route.route('/subtract/<int:item_id>/<int:quantity', methods=['PUT', 'DELETE'])
+def subtractDeleteItem(item_id, quantity):
 
-#While in PRODUCTS PAGE, add item to shopping cart
+    cart_item = db.session.query(CartItem).filter(CartItem.id == item_id).first()
+
+    if request.method == "DELETE":
+        try:
+            db.session.delete(cart_item)
+            db.session.commit()
+            return {'message': 'Cart item successfuly deleted'}, 200
+        except:
+            return {'errors': {'message': 'There was an error in deleting item from cart'}}
+
+    if request.method == "POST":
+        try:
+            cart_item.quantity = cart_item.quantity - quantity
+
+            if cart_item.quantity <= 0:
+                db.session.delete(cart_item)
+                db.session.commit()
+                return {'message': 'Cart item successfuly deleted'}, 200
+            else:
+                db.session.commit()
+                return {'cart_item': cart_item.to_dict()}
+        except:
+            return {'errors': {'message': 'There was an error in subtracting quantity from item'}}
+
+#Add item via PRODUCT id and CART id
 @cart_route.route('/<int:cart_id>/<int:product_id>', methods=['POST'])
 @login_required
 def addItem(cart_id, product_id):
@@ -88,7 +119,7 @@ def addItem(cart_id, product_id):
         item.quantity = item.quantity + 1
         try:
             db.session.commit()
-            return {'item': item.to_dict()}
+            return {'item': item.to_dict()}, 201
         except:
             return {'errors': {'message': 'Could not update quantity of item'}}
 
@@ -102,9 +133,83 @@ def addItem(cart_id, product_id):
         try:
             db.session.add(new_item)
             db.session.commit()
-            return {'new_item': new_item.to_dict()}
+            return {'new_item': new_item.to_dict()}, 201
         except:
             return {'errors': {'message': 'Could not update add item to cart'}}
 
+#Create new order
+@cart_route.route('/orders', methods=['POST'])
+@login_required
+def createOrder():
+    currentUser = current_user.to_dict()
+    user_id = currentUser['id']
+    order = Order(
+        user_id = user_id,
+        total = 0
+    )
+    try:
+        db.session.add(order)
+        db.session.commit()
+        return {'order': order.to_dict()}, 201
+    except:
+        return {'errors': {'message': 'There was an error creating order'}}
 
-#While in CART, Add/Subtract quantity of item
+#Add item to order
+@cart_route.route('/orders/<int:order_id>/<int:cart_item_id>')
+@login_required
+def addItemOrder(order_id, cart_item_id):
+
+    order = db.session.query(Order).filter(Order.id == order_id).first()
+    cart_item = db.session.query(CartItem).filter(CartItem.id == cart_item_id).first()
+
+    new_item = OrderItem(
+        order_id = order.id,
+        product_id = cart_item.product_id,
+        quantity = cart_item.quantity
+    )
+
+    try:
+        db.session.add(new_item)
+        db.session.commit()
+        return {'new_item': new_item.to_dict()}, 201
+    except:
+        return {'errors': {'message': 'There was an error in adding item to order'}}
+
+#GET Order items
+@cart_route.route('/orders/<int:order_id>')
+@login_required
+def getOrder(order_id):
+    order = db.session.query(Order).filter(Order.id == order_id).first()
+
+    if not order:
+        return {'errors': {'message': 'Cart does not exist'}}, 404
+
+    items = db.session.query(OrderItem).filter(OrderItem.order_id == order_id)
+    items_dict = [item.to_dict() for item in items]
+
+    # calculate cart total
+    total = 0
+
+    #get product info and images
+    for item in items_dict:
+        product = db.session.query(Product).filter(Product.id == item['product_id']).first()
+        images = db.session.query(ProductImage).filter(ProductImage.product_id == item['product_id'])
+
+        #add item cost to total
+        item_price = product.price * item['quantity']
+        total += item_price
+
+        #add product and images to item
+        item['product'] = product.to_dict()
+        item['images'] = [image.to_dict() for image in images]
+
+
+    #set total of cart
+    order.total = total
+    db.session.commit()
+
+    order_dict = order.to_dict()
+    order_dict['items'] = items_dict
+    order_dict['total'] = total
+
+    return {'order': order_dict}, 200
